+++ changes.md
@@ -11,6 +11,7 @@ Remove some low quality unmaintained devices, and add unit tests.
 * The legacy config_properties interface is deprecated.  It will still work for a while with the compatibility shims, but it raises warnings.
 * There is a fixed set of reserved keys.  No more device. prefixes.
 * Nonstandard additional properties added by the framework must go in a top level "extensions" section.
+* Closing subdevices is now the responsibilty of the main device, not the host.
 
 ## 0.24.0